HTTP errors management refactoring
