+++ app/view/error/403.twig.html
@@ -5,7 +5,7 @@
 {% endblock %}
 
 {% block headerImage %}
-	<div class="cover" style="background-image: url('{{ base_url }}public/img/header.jpg');"></div>
+	<div class="cover" style="background-image: url('{{ base_url }}public/img/error.gif'); background-position: top center;"></div>
 {% endblock %}
 
 {% block title %}
@@ -18,8 +18,10 @@
 
 {% block content %}
 	<div class="row">
-		<div class="col-xs-12">
-			Vous n'êtes pas autorisé(e) à consulter cette page.
+		<div class="col-xs-12 text-center">
+			<p>
+				Vous n'êtes pas autorisé(e) à consulter cette page.
+			</p>
 		</div>
 	</div>
 {% endblock %}
