+++ app/view/error/404.twig.html
@@ -5,7 +5,7 @@
 {% endblock %}
 
 {% block headerImage %}
-	<div class="cover" style="background-image: url('{{ base_url }}public/img/header.jpg');"></div>
+	<div class="cover" style="background-image: url('{{ base_url }}public/img/error.gif'); background-position: top center;"></div>
 {% endblock %}
 
 {% block title %}
@@ -18,8 +18,10 @@
 
 {% block content %}
 	<div class="row">
-		<div class="col-xs-12">
-			Une erreur s'est produite, la page recherchée n'existe pas :(
+		<div class="col-xs-12 text-center">
+			<p>
+				Une erreur s'est produite, la page recherchée n'existe pas.
+			</p>
 		</div>
 	</div>
 {% endblock %}
